commit before trying something new
the training method of the current umbc method can not match
the original attention sink method.

will try a 'sink finetuning' type method which finetunes a pretrained
model to work with a sink. If this works then we can look into adding a
UMBC pooler at the end of the current sequence.