ifixed score cache bug in cache
score cache was previously truncated to allow for a total
cache size which is a power of 2.

This created a mistake both in the positions which will be
applied to the last sub-cache, and the size of the scores
which cause illegal memory accesses.

The easiest solution was to undo the change, allow the cache size
to be a power of 2, and then require 64 sink tokens instead of
4, which allows the added sink tokens to work with flash attention
because 2**n + 64 will always fit into the flash attention block
size of 64 as long as 2**n >= 64

added test cases to verify that flash attention works with
the current size scheme

changed all relevant test cases to use 64 sinks
changed all test runner files to use 64 sinks